LayoutTests/inspector/debugger/symbolic-breakpoint-regex-case-sensitive-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for SymbolicBreakpoint.


== Running test suite: SymbolicBreakpoint
-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.Hit
Adding breakpoint...
Triggering breakpoint...
PASS: Should pause.

Removing breakpoint...
Triggering breakpoint...
PASS: Should not pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.Hit.Options.Condition

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.Hit.Options.Condition.ConsoleCommandLineAPI
Adding saved console value 'false'...

Setting condition to saved console value...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Adding saved console value 'true'...
Setting condition to saved console value...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.Hit.Options.IgnoreCount

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.Hit.Options.Action.Log

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.Hit.Options.Actions.Evaluate

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.Hit.Options.Actions.Evaluate.ConsoleCommandLineAPI
Adding saved console value '1'...

Adding evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '2'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '3'...
Editing evaluate action using saved console value...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Adding saved console value '4'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.Miss
Adding breakpoint...
Triggering breakpoint...
PASS: Should not pause.

Removing breakpoint...
Triggering breakpoint...
PASS: Should not pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.Disabled
Adding breakpoint...
Triggering breakpoint...
PASS: Should not pause.

Removing breakpoint...
Triggering breakpoint...
PASS: Should not pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.name.Original
Adding breakpoint...
Overriding name...
Triggering breakpoint...
PASS: Should pause.

Removing breakpoint...
Triggering breakpoint...
PASS: Should not pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.name.New
Adding breakpoint...
Overriding name...
Triggering breakpoint...
PASS: Should not pause.

Removing breakpoint...
Triggering breakpoint...
PASS: Should not pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.displayName.Original
Adding breakpoint...
Adding displayName...
Triggering breakpoint...
PASS: Should pause.

Removing displayName...
Triggering breakpoint...
PASS: Should pause.

Removing breakpoint...
Triggering breakpoint...
PASS: Should not pause.

-- Running test case: SymbolicBreakpoint.RegexCaseSensitive.displayName.New
Adding breakpoint...
Adding displayName...
Triggering breakpoint...
PASS: Should pause.

Removing displayName...
Triggering breakpoint...
PASS: Should not pause.

Removing breakpoint...
Triggering breakpoint...
PASS: Should not pause.
